pyproject.toml
[tool.poetry]
name = "crewai-multiagent-starter-v3.1"
version = "0.3.1"
description = "CrewAI multi-agent starter (LLM optional)."
authors = ["you <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = "^0.51.0"
jinja2 = "^3.1.4"
pydantic = "^2.8.0"
typer = "^0.12.3"
python-dotenv = "^1.0.1"
openai = "^1.40.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
